requirements.txt
Pillow==9.0.1
recordclass==0.17.2
